+++ frontend/imports/ui/client/headers/tabs.html
@@ -1,9 +1,9 @@
 <template name="tabs">
   <div class="tabs">
     <ul class="nav nav-tabs" role="tablist">
-      <li role="presentation" class="first-tab active"><a href="#trade" class="tab-link" aria-controls="trade" role="tab" data-toggle="tab">TRADE</a></li>
-      <li role="presentation" class="middle-tab"><a href="#deposit" class="tab-link" aria-controls="deposit" role="tab" data-toggle="tab">DEPOSIT / WITHDRAW</a></li>
-      <li role="presentation" class="last-tab"><a href="#transfer" class="tab-link" aria-controls="transfer" role="tab" data-toggle="tab">TRANSFER</a></li>
+      <li role="presentation" class="first-tab active" {{b "click: changeTab"}}><a href="#trade" class="tab-link" aria-controls="trade" role="tab" data-toggle="tab">TRADE</a></li>
+      <li role="presentation" class="middle-tab" {{b "click: changeTab"}}><a href="#deposit" class="tab-link" aria-controls="deposit" role="tab" data-toggle="tab">DEPOSIT / WITHDRAW</a></li>
+      <li role="presentation" class="last-tab" {{b "click: changeTab"}}><a href="#transfer" class="tab-link" aria-controls="transfer" role="tab" data-toggle="tab">TRANSFER</a></li>
     </ul>
   </div>
   <div class="vertical-line"></div>
